+++ app/controllers/admin/custom_template_nodes_controller.rb
@@ -1,0 +1,2 @@
+class Admin::CustomTemplateNodesController < Admin::ApplicationController
+end