+++ app/controllers/editorial/nodes_controller.rb
@@ -27,9 +27,7 @@ module Editorial
       configure_defaults!
       authorize! :create_in, @section
 
-      @node_types = Node.descendants.reject {|klass|
-        [RootNode, SectionHome].include? klass
-      }.map do |klass|
+      @node_types = [GeneralContent].map do |klass|
         name = klass.name.underscore
         [I18n.t("domain_model.nodes.#{name}"), name]
       end