+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This file is influenced by http://keepachangelog.com/.
 - Infrastructure and Telecommunications category (static)
 - Categories (static) shown on the homepage
 - Provided aggregate list of news from `/news`
+- Added a node type for nodes with custom templates
 
 ### Changed
 
@@ -20,6 +21,7 @@ This file is influenced by http://keepachangelog.com/.
 - News articles slugs are now a combination of release dates and article names
 - News articles can now be created without a parent
 - Nodes can now specify a layout to override with
+- Removed news article from the node creation prepare select box
 
 ### Fixed